refact(api): mvoing the controllers into the their specific files
