+++ apps/api/restaurants/controllers/__init__.py
@@ -2,8 +2,33 @@ from .favorites import (
     ReviewerFavoriteRestaurantListController,
     ReviewerFavoriteRestaurantController,
 )
+from .owner import OwnerDashboardController, OwnerRestaurantsController
+from .photos import (
+    RestaurantPhotoDetailController,
+    RestaurantPhotoPrimaryController,
+    RestaurantPhotosController,
+)
+from .restaurants import (
+    RestaurantDetailController,
+    RestaurantsController,
+)
+from .menu_items import (
+    RestaurantMenuItemDetailController,
+    RestaurantMenuItemsController,
+)
+from .homepage import RestaurantHomepageController
 
 __all__ = [
+    "RestaurantHomepageController",
     "ReviewerFavoriteRestaurantController",
     "ReviewerFavoriteRestaurantListController",
+    "OwnerDashboardController",
+    "OwnerRestaurantsController",
+    "RestaurantDetailController",
+    "RestaurantsController",
+    "RestaurantMenuItemDetailController",
+    "RestaurantMenuItemsController",
+    "RestaurantPhotoDetailController",
+    "RestaurantPhotoPrimaryController",
+    "RestaurantPhotosController",
 ]
